refactor(http): move ExtractTokenFromHeader from utils to http package for better organization
